Users/app/requirements.txt
aio-pika==8.2.2
aioredis==2.0.1
aiormq==6.4.2
alembic==1.8.1
anyio==3.6.1
async-timeout==4.0.2
click==8.1.3
Deprecated==1.2.13
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.85.0
greenlet==1.1.3
h11==0.13.0
hiredis==2.0.0
idna==3.4
Mako==1.2.2
MarkupSafe==2.1.1
multidict==6.0.2
packaging==21.3
pamqp==3.2.1
passlib==1.7.4
psycopg2==2.9.3
pydantic==1.10.2
PyJWT==2.5.0
pyparsing==3.0.9
python-dotenv==0.21.0
python-multipart==0.0.5
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
typing-extensions==4.3.0
uvicorn==0.18.3
wrapt==1.14.1
yarl==1.8.1
